requirements.txt
# Core ML/AI libraries
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Whisper AI
openai-whisper>=20230314

# Llama support
llama-cpp-python>=0.2.0
sentencepiece>=0.1.99

# Audio processing
soundfile>=0.12.0
librosa>=0.10.0
sounddevice>=0.4.6

# YouTube download
yt-dlp>=2023.10.0

# Graph RAG
nano-graphrag>=0.0.8
sentence-transformers>=2.2.0
networkx>=3.0

# Storage & Data models
pydantic>=2.0.0
pyyaml>=6.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# CLI (future)
typer>=0.9.0
rich>=13.0.0

# macOS Menu Bar App (Sprint 6) - macOS only
# rumps>=0.4.0  # Install manually on macOS: pip install rumps
# pynput>=1.7.0  # Install manually on macOS: pip install pynput
